+++ Changelog.md
@@ -5,6 +5,8 @@ Deprecations
 
 * Deprecate `be_xyz` predicate matcher when `xyz?` is a private method
   (Jon Rowe).
+* Deprecate `be_true` and `be_false` in favour of `be_truthy` and `be_falsey`.
+  (Sam Phippen)
 
 ### 2.14.0 / 2013-07-06
 [full changelog](http://github.com/rspec/rspec-expectations/compare/v2.14.0.rc1...v2.14.0)